requirements.txt
# Core dependencies
# Add your project dependencies here
google-adk
opentelemetry-instrumentation-google-genai
python-dotenv

# Development dependencies (optional)
# pytest==7.4.0
# black==23.7.0
# flake8==6.1.0
# mypy==1.5.0
